requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
cffi==1.13.2
chardet==3.0.4
coloredlogs==10.0
discord.py==1.2.4
humanfriendly==4.18
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pycparser==2.19
PyNaCl==1.3.0
python-dotenv==0.10.3
six==1.13.0
soupsieve==1.9.5
typing-extensions==3.7.4.1
urllib3==1.25.7
websockets==6.0
yarl==1.3.0
